added my adjusted collect files script
